btrfs-progs: move transaction.c to kernel-shared/
