+++ Android.mk
@@ -25,7 +25,7 @@ objects := kernel-shared/ctree.c kernel-shared/disk-io.c kernel-lib/radix-tree.c
           kernel-shared/ulist.c qgroup-verify.c backref.c common/string-table.c task-utils.c \
           kernel-shared/inode.c kernel-shared/file.c find-root.c \
 	  free-space-tree.c help.c cmds/receive-dump.c \
-          common/fsfeatures.c kernel-lib/tables.c kernel-lib/raid56.c transaction.c
+          common/fsfeatures.c kernel-lib/tables.c kernel-lib/raid56.c kernel-shared/transaction.c
 cmds_objects := cmds-subvolume.c cmds-filesystem.c cmds-device.c cmds-scrub.c \
                cmds-inspect.c cmds-balance.c cmds-send.c cmds-receive.c \
                cmds-quota.c cmds-qgroup.c cmds-replace.c cmds-check.c \
