+++ btrfs-list.c
@@ -27,7 +27,7 @@
 #include <dirent.h>
 #include <libgen.h>
 #include "kernel-shared/ctree.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "common/utils.h"
 #include "ioctl.h"
 #include <uuid/uuid.h>
