+++ btrfs-map-logical.c
@@ -26,7 +26,7 @@
 #include "volumes.h"
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "kernel-lib/list.h"
 #include "common/utils.h"
 #include "common/help.h"
