+++ check/mode-common.c
@@ -18,7 +18,7 @@
 #include "kernel-shared/ctree.h"
 #include "common/internal.h"
 #include "common/messages.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "common/utils.h"
 #include "kernel-shared/disk-io.h"
 #include "repair.h"
