+++ cmds/inspect-dump-tree.c
@@ -28,7 +28,7 @@
 #include "kernel-shared/ctree.h"
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "volumes.h"
 #include "cmds/commands.h"
 #include "common/utils.h"
