+++ cmds/rescue.c
@@ -21,7 +21,7 @@
 #include <getopt.h>
 #include "kernel-shared/ctree.h"
 #include "volumes.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "kernel-shared/disk-io.h"
 #include "cmds/commands.h"
 #include "common/utils.h"
