+++ common/utils.c
@@ -46,7 +46,7 @@
 #include "kernel-lib/radix-tree.h"
 #include "kernel-shared/ctree.h"
 #include "kernel-shared/disk-io.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "crypto/crc32c.h"
 #include "common/utils.h"
 #include "common/path-utils.h"
