+++ convert/main.c
@@ -94,7 +94,7 @@
 #include "kernel-shared/ctree.h"
 #include "kernel-shared/disk-io.h"
 #include "volumes.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "common/utils.h"
 #include "common/task-utils.h"
 #include "common/path-utils.h"
