+++ convert/source-reiserfs.c
@@ -23,7 +23,7 @@
 #include <sys/stat.h>
 #include <stdbool.h>
 #include "kernel-shared/disk-io.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "common/utils.h"
 #include "kernel-lib/bitops.h"
 #include "convert/common.h"
