+++ kernel-shared/backref.c
@@ -21,7 +21,7 @@
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/backref.h"
 #include "kernel-shared/ulist.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "common/internal.h"
 
 #define pr_debug(...) do { } while (0)
