+++ kernel-shared/delayed-ref.c
@@ -6,7 +6,7 @@
 #include "kernel-shared/ctree.h"
 #include "btrfs-list.h"
 #include "kernel-shared/delayed-ref.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 
 /*
  * delayed back reference update tracking.  For subvolume trees
