+++ kernel-shared/free-space-tree.c
@@ -21,7 +21,7 @@
 #include "kernel-shared/free-space-cache.h"
 #include "kernel-shared/free-space-tree.h"
 #include "volumes.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "kernel-lib/bitops.h"
 #include "common/internal.h"
