+++ kernel-shared/transaction.h
