+++ kernel-shared/uuid-tree.c
@@ -20,7 +20,7 @@
 #include <uuid/uuid.h>
 #include <sys/ioctl.h>
 #include "kernel-shared/ctree.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 #include "common/utils.h"
