+++ mkfs/main.c
@@ -35,7 +35,7 @@
 #include "kernel-shared/ctree.h"
 #include "kernel-shared/disk-io.h"
 #include "volumes.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 #include "common/utils.h"
 #include "common/path-utils.h"
 #include "common/device-utils.h"
