+++ quick-test.c
@@ -24,7 +24,7 @@
 #include "kernel-shared/ctree.h"
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
-#include "transaction.h"
+#include "kernel-shared/transaction.h"
 
 /* for testing only */
 static int next_key(int i, int max_key) {
